test(hygiene): colocated tests move to the documented __tests__ mirror
CONTRIBUTING's "Project layout" says tests live under src/__tests__/
mirroring the source path. Thirteen test files had drifted to sit next
to their subject instead (two in i18n/, two in orchestrator/, eight in
services/, one in tools/). This moves each to the mirror path and
adjusts only what depends on file depth: relative import specifiers
(./x.js -> ../../<dir>/x.js, ../x.js -> ../../x.js) and the two
import.meta.dirname walks in the i18n tests that locate the repo root.
No test body, name, or assertion changes.

Two consequences worth knowing. The tarball shrinks: tsconfig excludes
only src/__tests__, so these thirteen were being compiled into dist/
and shipped with their source maps (npm pack --dry-run listed 26
.test.* entries before, 0 after). And for the same reason they are no
longer type-checked by `npm run lint`, matching every other test in the
repo; they still run under vitest, whose include glob already covered
both locations, and the collected file count is unchanged at 605.
